Create a seperate profile for HyperWalker
spawnFf() needs to be cleaned up. I'm not sure if this even works as
intended. We want to be sure that the profile creation happens _before_
we spawn the Firefox marionette.

One nice thing is that the profile creation command appears to exit
nicely without causing any problem if the profile already exists.