Refactoring and implementation of several new features
HyperWalker is now able to "surf the web". When a link is clicked on a
page, another snapshot is created and then loaded in a new webview.

The page title is now added to the saved snapshot file.

Quite a few more fixes, refactors, and improvements. However, it's a bit
more messy than before in certain places, and the comment ration has
decreased. Will make headway here in upcoming commits.